buggy

+++ firmware/lib/serial/src/serial.c
@@ -102,8 +102,8 @@ void
 serial_pkg_init(void)
 {
     #if MYNEWT_VAL(UWB_PKG_INIT_LOG)
-    printf("{\"utime\": %"PRIu32",\"msg\": \"button_pkg_init\"}\n",
-           (uint32_t)dpl_cputime_ticks_to_usecs(dpl_cputime_get32()));
+    printf("{\"utime\": %lu,\"msg\": \"button_pkg_init\"}\n",
+           dpl_cputime_ticks_to_usecs(dpl_cputime_get32()));
     #endif
 
     rbuf_init(&g_serial.rbuf_tx, g_serial.buffer_tx, MYNEWT_VAL(SERIAL_TX_RBUF_SIZE));
